+++ poc/.clinerules
