+++ poc/SplitTheBillPoc.Tests/SplitTheBillPoc.Tests.csproj
