+++ poc/SplitTheBillPoc.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SplitTheBillPoc", "SplitTheBillPoc\SplitTheBillPoc.csproj", "{45CA63C2-A624-49D2-A925-95E476D5AFB6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SplitTheBillPoc", "SplitTheBillPoc\SplitTheBillPoc.csproj", "{4E48D5C6-3FDF-4D5E-AF25-F9C493D4E515}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SplitTheBillPoc.Tests", "SplitTheBillPoc.Tests\SplitTheBillPoc.Tests.csproj", "{E0FE038D-97E7-495C-AE03-AC9F283D20A1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -14,9 +16,13 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{45CA63C2-A624-49D2-A925-95E476D5AFB6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{45CA63C2-A624-49D2-A925-95E476D5AFB6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{45CA63C2-A624-49D2-A925-95E476D5AFB6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{45CA63C2-A624-49D2-A925-95E476D5AFB6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4E48D5C6-3FDF-4D5E-AF25-F9C493D4E515}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4E48D5C6-3FDF-4D5E-AF25-F9C493D4E515}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4E48D5C6-3FDF-4D5E-AF25-F9C493D4E515}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4E48D5C6-3FDF-4D5E-AF25-F9C493D4E515}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E0FE038D-97E7-495C-AE03-AC9F283D20A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0FE038D-97E7-495C-AE03-AC9F283D20A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0FE038D-97E7-495C-AE03-AC9F283D20A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0FE038D-97E7-495C-AE03-AC9F283D20A1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
