+++ poc/SplitTheBillPoc/Data/EntityTypeConfigurations/ExpenseConfiguration.cs
