+++ poc/SplitTheBillPoc/Data/EntityTypeConfigurations/GroupConfiguration.cs
