+++ poc/SplitTheBillPoc/Data/EntityTypeConfigurations/GroupMemberConfiguration.cs
