+++ poc/SplitTheBillPoc/Data/EntityTypeConfigurations/MemberConfiguration.cs
