+++ poc/SplitTheBillPoc/Data/EntityTypeConfigurations/PaymentConfiguration.cs
