+++ poc/SplitTheBillPoc/Models/Expense.cs
@@ -2,9 +2,9 @@ namespace SplitTheBillPoc.Models;
 
 internal sealed class Expense
 {
-    public Guid Id { get; set; }
-    public required decimal Amount { get; set; }
+    public required Guid Id { get; set; }
+    public required Guid GroupId { get; set; }
     
-    public required Guid GroupId { get; init; }
-    public required Guid PaidByMemberId { get; set; }
+    public required string Description { get; set; }
+    public required decimal Amount { get; set; }
 }
