+++ poc/SplitTheBillPoc/Models/Group.cs
@@ -2,10 +2,10 @@ namespace SplitTheBillPoc.Models;
 
 internal sealed class Group
 {
-    public Guid Id { get; init; }
+    public required Guid Id { get; init; }
     public required string Name { get; set; }
 
-    public ICollection<Member> Members { get; set; } = [];
-    public ICollection<Expense> Expenses { get; set; } = [];
-    public ICollection<Payment> Payments { get; set; } = [];
+    public List<Member> Members { get; set; } = [];
+    public List<Expense> Expenses { get; set; } = [];
+    public List<Payment> Payments { get; set; } = [];
 }
