+++ poc/SplitTheBillPoc/Models/GroupMember.cs
