+++ poc/SplitTheBillPoc/Models/Member.cs
@@ -1,7 +1,9 @@
 namespace SplitTheBillPoc.Models;
 
-public sealed class Member
+internal sealed class Member
 {
-    public Guid Id { get; init; }
+    public required Guid Id { get; init; }
     public required string Name { get; set; }
+
+    public List<Group> Groups { get; set; } = [];
 }
