+++ poc/SplitTheBillPoc/Program.cs
@@ -1,7 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using SplitTheBillPoc.Data;
 using SplitTheBillPoc.Modules.Groups;
-using SplitTheBillPoc.Modules.Members;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -11,9 +10,5 @@ builder.Services
     );
 
 var app = builder.Build();
-
-app
-    .MapMembersModule()
-    .MapGroupsModule();
-
+app.MapGroupsModule();
 app.Run();
