+++ poc/design-and-requirements.md
