--- poc2/SplitTheBillPoc/Models/Member.cs
@@ -1,9 +1,0 @@
-namespace SplitTheBillPoc.Models;
-
-internal sealed class Member
-{
-    public required Guid Id { get; init; }
-    public required string Name { get; set; }
-
-    public List<Group> Groups { get; set; } = [];
-}
