--- poc2/SplitTheBillPoc/Program.cs
@@ -1,14 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using SplitTheBillPoc.Data;
-using SplitTheBillPoc.Modules.Groups;
-
-var builder = WebApplication.CreateBuilder(args);
-
-builder.Services
-    .AddDbContext<AppDbContext>(optionsBuilder => optionsBuilder
-        .UseNpgsql(builder.Configuration.GetConnectionString("AppDbContext"))
-    );
-
-var app = builder.Build();
-app.MapGroupsModule();
-app.Run();
